Add assembly coverage information and improve result summary
This commit adds the recording and storing of assembly coverage information (mean and standard deviation) obtained from the SPAdes assembly log file. This information is then written to the PlasmidSeqRun model in our database.

The assembly statistics (mean coverage and standard deviation) are added to each job and written to the result summary spreadsheet. This enhancement provides more detailed assembly information and clarifies assembly statistics, facilitating a better understanding of the assembly process and its accuracy.

Moreover, the commit revises the structure of the result summary, providing a more intuitive and cleaner overview. Also, potential errors in the processing pipeline were handled more explicitly to avoid breaking the process and allowing its correct completion.

This improvement to our pipeline contributes to more clear and comprehensive results, enabling easier interpretation of our NGS data.

Signed-off-by: Robert Warden-Rothman <rob.wr@grobio.com>