Refactor sequence match counts calculation, optimize output data writing
In this revision, the calculation of forward and reverse sequence match counts has been moved to inside assembly item creation loop. This is a logical optimization as these calculations should be specific to each assembly.

Additionally, the writing of output data to the excel file has been enhanced. The 'assembly_data_df', 'feature_data_df' and 'poly_data_df' are now created and populated ahead, improving the readability of the code. The dataframe 'assembly_data_df' is also extended with 'del_feature_count', 'mod_feature_count' and 'indel_count' for better insights.

These changes should improve the overall efficiency and readability of the code.

Signed-off-by: Robert Warden-Rothman <rob.wr@grobio.com>